scripts/requirements.txt
colour-science==0.4.7
numpy==2.3.5
